+++ mlir/include/mlir/Support/LLVM.h
@@ -99,6 +99,7 @@ namespace mlir {
 using llvm::cast;
 using llvm::cast_or_null;
 using llvm::dyn_cast;
+using llvm::dyn_cast_if_present;
 using llvm::dyn_cast_or_null;
 using llvm::isa;
 using llvm::isa_and_nonnull;